+++ packages/core/src/api/logs/const.ts
@@ -1,0 +1,12 @@
+import type { LogArgsSerializer } from './types';
+
+export const defaultLogArgsSerializer: LogArgsSerializer = (args) =>
+  args
+    .map((arg) => {
+      try {
+        return String(arg);
+      } catch (err) {
+        return '';
+      }
+    })
+    .join(' ');