+++ packages/core/src/api/logs/index.ts
@@ -1,3 +1,5 @@
+export { defaultLogArgsSerializer } from './const';
+
 export { initializeLogsAPI } from './initialize';
 
 export type { LogContext, LogEvent, LogArgsSerializer, LogsAPI, PushLogOptions } from './types';